requirements.txt
aiogram==3.10.0
requests==2.32.3
python-dotenv
beautifulsoup4
selenium